rearranging necessary functions
